IZPACK-1464: UserInputPanel - Search field issues
