+++ izpack-panel/src/main/java/com/izforge/izpack/panels/userinput/console/ConsoleFieldFactory.java
@@ -32,10 +32,10 @@ import com.izforge.izpack.panels.userinput.console.custom.ConsoleCustomField;
 import com.izforge.izpack.panels.userinput.console.divider.ConsoleDividerField;
 import com.izforge.izpack.panels.userinput.console.file.ConsoleDirField;
 import com.izforge.izpack.panels.userinput.console.file.ConsoleFileField;
-import com.izforge.izpack.panels.userinput.console.file.ConsoleSearchField;
 import com.izforge.izpack.panels.userinput.console.password.ConsolePasswordGroupField;
 import com.izforge.izpack.panels.userinput.console.radio.ConsoleRadioField;
 import com.izforge.izpack.panels.userinput.console.rule.ConsoleRuleField;
+import com.izforge.izpack.panels.userinput.console.search.ConsoleSearchField;
 import com.izforge.izpack.panels.userinput.console.spacer.ConsoleSpacerField;
 import com.izforge.izpack.panels.userinput.console.staticText.ConsoleStaticText;
 import com.izforge.izpack.panels.userinput.console.text.ConsoleTextField;
